added encryption
